refactor: rewrite core logic for improved readability and performance
Refactor Bluetooth and Wi-Fi handling for improved modularity
Replaced outdated and redundant code with a more modular architecture for Bluetooth and Wi-Fi operations.
Introduced `LCDHandler` and `PreferencesHandler` for better separation of concerns and code reuse.
Removed legacy matrix handling, EEPROM dependencies, and unused utilities to streamline functionality.

+++ src/utils/rtc_handler.cpp
@@ -6,7 +6,7 @@ RTCHandler::RTCHandler(const uint8_t resetPin, const uint8_t clockPin, const uin
 
 void RTCHandler::initialize() {
     rtc.init();
-    
+
     if (rtc.isHalted()) {
         Serial.println(F("RTC was not running, setting default time"));
         Ds1302::DateTime dt = {
@@ -32,13 +32,13 @@ String RTCHandler::getFormattedTime() {
     const Ds1302::DateTime now = getCurrentDateTime();
 
     String timeStr = "";
-    
+
     if (now.hour < 10) timeStr += "0";
     timeStr += String(now.hour) + ":";
-    
+
     if (now.minute < 10) timeStr += "0";
     timeStr += String(now.minute) + ":";
-    
+
     if (now.second < 10) timeStr += "0";
     timeStr += String(now.second);
 
@@ -47,15 +47,15 @@ String RTCHandler::getFormattedTime() {
 
 String RTCHandler::getFormattedDate() {
     const Ds1302::DateTime now = getCurrentDateTime();
-    const String months[] = { "Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec" };
-    
+    const String months[] = {"Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Oct", "Nov", "Dec"};
+
     String dateStr = "";
-    
+
     if (now.day < 10) dateStr += "0";
     dateStr += String(now.day) + " ";
     dateStr += months[now.month - 1] + " ";
-    dateStr += "20" + String(now.year);  // Assuming 20xx years
-    
+    dateStr += "20" + String(now.year); // Assuming 20xx years
+
     return dateStr;
 }
 
@@ -65,7 +65,7 @@ String RTCHandler::getFormattedDateTime() {
 
 String RTCHandler::getFormattedWeekDay() {
     const Ds1302::DateTime now = getCurrentDateTime();
-    static String const daysOfWeek[] = { "Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun" };
+    static String const daysOfWeek[] = {"Mon", "Tue", "Wed", "Thu", "Fri", "Sat", "Sun"};
     return daysOfWeek[now.dow - 1];
 }
 
@@ -75,51 +75,51 @@ void RTCHandler::setTimeFromEpoch(const unsigned long epochSeconds) {
     constexpr unsigned long secondsPerDay = 86400;
     constexpr unsigned long secondsPerYear = 365 * secondsPerDay;
     constexpr unsigned long secondsPerLeapYear = 366 * secondsPerDay;
-    
+
     int year = 1970;
     unsigned long remainingSeconds = epochSeconds;
-    
+
     while (true) {
         const bool isLeapYear = year % 4 == 0 && (year % 100 != 0 || year % 400 == 0);
         const unsigned long yearSeconds = isLeapYear ? secondsPerLeapYear : secondsPerYear;
-        
+
         if (remainingSeconds < yearSeconds) {
             break;
         }
-        
+
         remainingSeconds -= yearSeconds;
         year++;
     }
-    
+
     // Calculate month
     int daysInMonth[] = {31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
     if (year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)) {
         daysInMonth[1] = 29; // February has 29 days in leap years
     }
-    
+
     int month = 0;
     unsigned long days = remainingSeconds / secondsPerDay;
     remainingSeconds %= secondsPerDay;
-    
+
     while (days >= daysInMonth[month]) {
         days -= daysInMonth[month];
         month++;
     }
-    
+
     // Calculate day, hour, minute, second
-    const int day = days + 1; // Days are 1-based
-    const int hour = remainingSeconds / 3600;
+    const int day = static_cast<int>(days + 1); // Days are 1-based
+    const int hour = static_cast<int>(remainingSeconds / 3600);
     remainingSeconds %= 3600;
-    const int minute = remainingSeconds / 60;
-    const int second = remainingSeconds % 60;
-    
+    const int minute = static_cast<int>(remainingSeconds / 60);
+    const int second = static_cast<int>(remainingSeconds % 60);
+
     // Calculate day of the week (0 = Sunday, 1 = Monday, ..., 6 = Saturday)
     // January 1, 1970, was a Thursday (4)
     const unsigned long totalDays = epochSeconds / secondsPerDay;
-    int dow = (totalDays + 4) % 7;
+    int dow = static_cast<int>((totalDays + 4) % 7);
     // Convert to DS1302 format where 1 = Monday, ..., 7 = Sunday
     dow = dow == 0 ? 7 : dow;
-    
+
     // Set the RTC
     Ds1302::DateTime dt = {
         .year = static_cast<uint8_t>(year % 100), // 2-digit year
@@ -130,17 +130,17 @@ void RTCHandler::setTimeFromEpoch(const unsigned long epochSeconds) {
         .second = static_cast<uint8_t>(second),
         .dow = static_cast<uint8_t>(dow)
     };
-    
+
     rtc.setDateTime(&dt);
 }
 
 unsigned long RTCHandler::getEpochTime() {
     const Ds1302::DateTime now = getCurrentDateTime();
-    
+
     // Convert to Unix timestamp
     // This is a simplified calculation that doesn't account for all leap years correctly
     const int year = 2000 + now.year; // Assuming 20xx years
-    
+
     // Count seconds for years since 1970
     unsigned long seconds = 0;
     for (int y = 1970; y < year; y++) {
@@ -150,23 +150,23 @@ unsigned long RTCHandler::getEpochTime() {
             seconds += 365 * 24 * 60 * 60; // Regular year
         }
     }
-    
+
     // Add seconds for the months in the current year
     int daysInMonth[] = {31, 28, 31, 30, 31, 30, 31, 31, 30, 31, 30, 31};
     if (year % 4 == 0 && (year % 100 != 0 || year % 400 == 0)) {
         daysInMonth[1] = 29; // February has 29 days in leap years
     }
-    
+
     for (int m = 0; m < now.month - 1; m++) {
         seconds += daysInMonth[m] * 24 * 60 * 60;
     }
-    
+
     // Add seconds for days, hours, minutes, seconds
     seconds += (now.day - 1) * 24 * 60 * 60;
     seconds += now.hour * 60 * 60;
     seconds += now.minute * 60;
     seconds += now.second;
-    
+
     return seconds;
 }
